requirements.txt
HARPy==0.2.0
browsermob-proxy==0.5.0
execnet==1.1
py==1.4.15
py.saunter==0.63
pytest==2.3.5
pytest-marks==0.4
pytest-xdist==1.8
requests==1.2.3
selenium==2.35.0
wsgiref==0.1.2
-e git+git@github.com:OneOcean/clipcard-python-sdk.git@8d38fe1e4d4542ce026762a0367624dc402ccf5a#egg=clipcard_client-dev
